Move duplicate method to the utils
