+++ q2_types/genome_data/tests/test_methods.py
@@ -21,9 +21,10 @@ from q2_types.genome_data import (
     LociDirectoryFormat
 )
 from q2_types.genome_data._methods import (
-    collate_loci, collate_genomes, _duplicate_with_warning, _collate_helper,
+    collate_loci, collate_genomes, _collate_helper,
     collate_orthologs, partition_orthologs, collate_genes, collate_proteins
 )
+from q2_types._util import _duplicate_with_warning
 
 
 class TestPartitionCollating(TestPluginBase):
@@ -121,19 +122,6 @@ class TestPartitionCollating(TestPluginBase):
         collated_loci = collate_loci(loci_list)
         self.assertTrue(all(os.path.exists(
             collated_loci.path / f"loci{no}.gff") for no in [1, 2, 3, 4]))
-
-    def test_duplicate_warning(self):
-        tmpdir = self.temp_dir.name
-        src = os.path.join(tmpdir, "file.txt")
-        dst = os.path.join(tmpdir, "file_copy.txt")
-        with open(src, "w"), open(dst, "w"):
-            pass
-
-        with warnings.catch_warnings(record=True) as w:
-            warnings.simplefilter("always")
-            _duplicate_with_warning(src, dst)
-
-            self.assertIn("File already exists", str(w[-1].message))
 
     def test_partition_orthologs(self):
         p = self.get_data_path("collated_orthologs")
